+++ packages/evolution-common/src/services/questionnaire/__tests__/index.test.ts
@@ -72,6 +72,7 @@ describe('QuestionnaireFactory', () => {
             'segmentSameModeAsReverseTrip',
             'segmentModePre',
             'segmentMode',
+            'segmentModeOtherSpecify',
             'segmentHasNextMode',
             'personVisitedPlacesMap',
             'buttonConfirmNextSection'